# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RemindMe Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "200":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{uid}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List reminders for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    get:
      operationId: getReminder
      tags: [reminders]
      summary: Get reminder by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReminderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [reminders]
      summary: Get dashboard summary for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        displayName:
          type: string
        email:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - uid
        - displayName
        - email
        - createdAt
    RegisterUserBody:
      type: object
      properties:
        uid:
          type: string
        displayName:
          type: string
        email:
          type: string
        photoUrl:
          type: ["string", "null"]
      required:
        - uid
        - displayName
        - email
    Reminder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [reminder, note, alarm]
        scheduledAt:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - content
        - type
        - isCompleted
        - priority
        - createdAt
        - updatedAt
    CreateReminderBody:
      type: object
      properties:
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [reminder, note, alarm]
        scheduledAt:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
      required:
        - userId
        - title
        - content
        - type
    UpdateReminderBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        scheduledAt:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
...
    DashboardSummary:
      type: object
      properties:
        totalReminders:
          type: integer
        pendingReminders:
          type: integer
        completedReminders:
          type: integer
        totalAlarms:
          type: integer
        activeAlarms:
          type: integer
        totalGroups:
          type: integer
        upcomingCount:
          type: integer
      required:
        - totalReminders
        - pendingReminders
        - completedReminders
        - totalAlarms
        - activeAlarms
        - totalGroups
        - upcomingCount