# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalReminders: number;
  pendingReminders: number;
  completedReminders: number;
  totalAlarms: number;
  activeAlarms: number;
  totalGroups: number;
  upcomingCount: number;
}