# lib/api-zod/src/generated/types/registerUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterUserBody {
  uid: string;
  displayName: string;
  email: string;
  /** @nullable */
  photoUrl?: string | null;
}