# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderPriority } from "./reminderPriority";
import type { ReminderType } from "./reminderType";

export interface Reminder {
  id: number;
  userId: string;
  title: string;
  content: string;
  type: ReminderType;
  /** @nullable */
  scheduledAt?: string | null;
  isCompleted: boolean;
  priority: ReminderPriority;
  createdAt: string;
  updatedAt: string;
}